requirements.txt
python-telegram-bot==21.3
flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
boto3==1.34.144
watchdog==4.0.1
psycopg2-binary==2.9.9; python_version < "3.14"
